+++ src/pifmrds/pifmrds.cpp
@@ -245,7 +245,8 @@ namespace pifmrds {
         proc.encoder().setPs(params.ps);
         proc.encoder().setRt(params.rt);
 
-        ngfmdmasync dma{params.transmissionFrequency, MPX_SAMPLE_RATE, DMA_BIT_DEPTH, DMA_FIFO_SIZE};
+        ngfmdmasync dma{params.transmissionFrequency, static_cast<uint32_t>(MPX_SAMPLE_RATE), DMA_BIT_DEPTH,
+                        DMA_FIFO_SIZE};
 
         // AudioPipeline owns source-rate input buffers, loop-aware EOF handling,
         // channel preservation, and source -> 228 kHz rate conversion. The FM-RDS
